+++ mhtmlextractor/__init__.py
@@ -1,0 +1,6 @@
+"""Public package interface for MHTML extraction."""
+
+from .extractor import MHTMLExtractor
+from .stats import ExtractionStats
+
+__all__ = ["ExtractionStats", "MHTMLExtractor"]
